Time: 3 ms (75.69%), Space: 69.5 MB (62.39%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -1,17 +1,17 @@
 class Solution {
     public double findMaxAverage(int[] nums, int k) {
-        double ans=0.0;
+        int ans=0;
         for(int i=0;i<k;i++){
             ans+=nums[i];
         }
         int r=k-1;
-        double maxavg=ans;
+        int maxavg=ans;
         while(r<nums.length-1){
             r++;
             ans=ans-nums[r-k]+nums[r];
             maxavg=Math.max(maxavg,ans);
         }
-        return maxavg/k;
+        return (double)maxavg/k;
 
     }
 }
